Update Robocopy-Mirror.ps1
Added default log location to path that the script is running from. Unless that location is the templocation for the temp script then it logs to the user desktop.
Added a variable to control if the script should try and elevate to run as admin.
Added the script now detects if it is running as admin and adjusts backup options to use /ZB if so.
Added basic sanity checks for source and destination. Will now popup with error if not sufficient.
Added runtime banner.
Fixed bug when checking if using a temp script.